refactor: move example code to example
