+++ src/jax_mask_efficient.py
@@ -160,218 +160,3 @@ def model_evaluation(state, test_data, test_labels):
         accs.append(eval_fn(state, pb, yb))
 
     return np.mean(np.array(accs))
-
-
-## Main Loop
-
-
-def main(args):
-
-    jax.clear_caches()
-
-    print(args, flush=True)
-
-    train_images, train_labels, test_images, test_labels = import_data_efficient_mask()
-
-    steps = args.epochs * math.ceil(len(train_images) / args.bs)
-
-    q = 1 / math.ceil(len(train_images) / args.bs)
-
-    noise_std = calculate_noise(q, args.epsilon, args.target_delta, args.epochs, args.accountant)
-    C = args.grad_norm
-
-    config = namedtuple("Config", ["momentum", "learning_rate"])
-    config.momentum = 1
-    config.learning_rate = args.lr
-
-    num_classes = args.num_classes
-
-    state = create_train_state(
-        model_name=args.model,
-        num_classes=num_classes,
-        config=config,
-    )
-
-    orig_dimension = 32
-    full_data_size = train_images.shape[0]
-    physical_bs = args.phy_bs
-    num_iter = steps
-
-    times = []
-    logical_batch_sizes = []
-
-    splits_test = jnp.split(test_images, 10)
-    splits_labels = jnp.split(test_labels, 10)
-    private = False
-
-    # Check privacy
-    if args.clipping_mode == "DP":
-        private = True
-
-    if private:
-
-        @jax.jit
-        def body_fun(t, args):
-            (
-                state,
-                accumulated_clipped_grads,
-                logical_batch_X,
-                logical_batch_y,
-                masks,
-            ) = args
-            # slice
-            start_idx = t * physical_bs
-            pb = jax.lax.dynamic_slice(
-                logical_batch_X,
-                (start_idx, 0, 0, 0, 0),
-                (physical_bs, 1, 3, orig_dimension, orig_dimension),
-            )
-            yb = jax.lax.dynamic_slice(logical_batch_y, (start_idx,), (physical_bs,))
-            mask = jax.lax.dynamic_slice(masks, (start_idx,), (physical_bs,))
-
-            # compute grads and clip
-            per_example_gradients = compute_per_example_gradients(state, pb, yb)
-            sum_of_clipped_grads_from_pb = process_a_physical_batch(per_example_gradients, mask, C)
-            accumulated_clipped_grads = add_trees(accumulated_clipped_grads, sum_of_clipped_grads_from_pb)
-
-            return (
-                state,
-                accumulated_clipped_grads,
-                logical_batch_X,
-                logical_batch_y,
-                masks,
-            )
-
-    else:
-
-        @jax.jit
-        def body_fun(t, args):
-            (
-                state,
-                accumulated_clipped_grads,
-                logical_batch_X,
-                logical_batch_y,
-                masks,
-            ) = args
-            # slice
-            start_idx = t * physical_bs
-            pb = jax.lax.dynamic_slice(
-                logical_batch_X,
-                (start_idx, 0, 0, 0),
-                (physical_bs, 3, orig_dimension, orig_dimension),
-            )
-            yb = jax.lax.dynamic_slice(logical_batch_y, (start_idx,), (physical_bs,))
-            mask = jax.lax.dynamic_slice(masks, (start_idx,), (physical_bs,))
-
-            # compute grads and clip
-            sum_of_grads = compute_gradients_non_dp(state, pb, yb, mask)
-            accumulated_grads = add_trees(accumulated_clipped_grads, sum_of_grads)
-
-            return state, accumulated_grads, logical_batch_X, logical_batch_y, masks
-
-    for t in range(num_iter):
-        sampling_rng = jax.random.PRNGKey(t + 1)
-        batch_rng, binomial_rng, noise_rng = jax.random.split(sampling_rng, 3)
-
-        #######
-        # poisson subsample
-        actual_batch_size = jax.device_put(
-            jax.random.bernoulli(binomial_rng, shape=(full_data_size,), p=q).sum(),
-            jax.devices("cpu")[0],
-        )
-        n_physical_batches = actual_batch_size // physical_bs + 1
-        logical_batch_size = n_physical_batches * physical_bs
-        n_masked_elements = logical_batch_size - actual_batch_size
-
-        # take the logical batch
-        indices = jax.random.permutation(batch_rng, full_data_size)[:logical_batch_size]
-        logical_batch_X = train_images[indices]
-        if private:
-            logical_batch_X = logical_batch_X.reshape(-1, 1, 3, orig_dimension, orig_dimension)
-        logical_batch_y = train_labels[indices]
-        #######
-
-        # masks
-        masks = jax.device_put(
-            jnp.concatenate([jnp.ones(actual_batch_size), jnp.zeros(n_masked_elements)]),
-            jax.devices("cpu")[0],
-        )
-
-        # cast to GPU
-        logical_batch_X = jax.device_put(logical_batch_X, jax.devices("gpu")[0])
-        logical_batch_y = jax.device_put(logical_batch_y, jax.devices("gpu")[0])
-        masks = jax.device_put(masks, jax.devices("gpu")[0])
-
-        print("##### Starting gradient accumulation #####", flush=True)
-        ### gradient accumulation
-        params = state.params
-
-        accumulated_clipped_grads0 = jax.tree.map(lambda x: 0.0 * x, params)
-
-        start = time.time()
-
-        # Main loop
-        if private:
-            _, accumulated_clipped_grads, *_ = jax.lax.fori_loop(
-                0,
-                n_physical_batches,
-                body_fun,
-                (
-                    state,
-                    accumulated_clipped_grads0,
-                    logical_batch_X,
-                    logical_batch_y,
-                    masks,
-                ),
-            )
-            noisy_grad = noise_addition(noise_rng, accumulated_clipped_grads, noise_std, C)
-            # update
-            state = jax.block_until_ready(update_model(state, noisy_grad))
-        else:
-            _, accumulated_grads, *_ = jax.lax.fori_loop(
-                0,
-                n_physical_batches,
-                body_fun,
-                (
-                    state,
-                    accumulated_clipped_grads0,
-                    logical_batch_X,
-                    logical_batch_y,
-                    masks,
-                ),
-            )
-            # update
-            state = jax.block_until_ready(update_model(state, accumulated_grads))
-
-        end = time.time()
-        duration = end - start
-
-        times.append(duration)
-        logical_batch_sizes.append(logical_batch_size)
-
-        print(logical_batch_size / duration, flush=True)
-
-        acc_iter = model_evaluation(state, splits_test, splits_labels)
-        print("iteration", t, "acc", acc_iter, flush=True)
-
-        # Compute privacy guarantees
-        if private:
-            epsilon, delta = compute_epsilon(
-                steps=t + 1,
-                batch_size=actual_batch_size,
-                num_examples=len(train_images),
-                target_delta=args.target_delta,
-                noise_multiplier=noise_std,
-            )
-            privacy_results = {"eps_rdp": epsilon, "delta_rdp": delta}
-            print(privacy_results, flush=True)
-
-    acc_last = model_evaluation(state, splits_test, splits_labels)
-
-    print("times \n", times, flush=True)
-
-    print("batch sizes \n ", logical_batch_size, flush=True)
-
-    print("accuracy at end of training", acc_last, flush=True)
-    thr = np.mean(np.array(logical_batch_sizes) / np.array(times))
-    return thr, acc_last
